requirements.txt
# Amul Product Alert Bot - Dependencies
# Install with: pip install -r requirements.txt

# Telegram Bot Framework
python-telegram-bot>=20.0

# Web Scraping
selenium>=4.15.0
requests>=2.31.0

# Database
psycopg2-binary>=2.9.9

# Environment Variables
python-dotenv>=1.0.0
